flask-app/requirements.txt
Flask
requests
python-jose
